fix wrongful reassembly based on SDU sequence number reset on handover
If the sending node is facing multiple handovers the SDU sequence number
is reset quickly which may lead to a reassembly of packet part which
share the same SDU sequence number but which are actually different packets.
Comparing the PduSno, which must be directly following in the UM mode, this
case can be checked.