MacNodeId: define using a "strong typedef"-like construct in LteCommon.h
Motivation: we want to ensure that MacNodeIds are not accidentally mixed
up with (assigned to/from) other types of IDs. A plain typedef does not
do this, as it allows implicit conversion from/to integers (and other
typedefs). In that regard, it is not very different from a #define.

C++ has no built-in support for "strong typedefs", but it can be
emulated in multiple ways, e.g. by wrapping the int into a class.
However, that would be too clumsy and an overkill for our purposes.
Instead, we (mis)use an "enum class" for that purpose. An enum
class cannot be implicitly converted to/from numeric types or other
enums, so it does the job. The codebase contains a few places where
MacNodeId is used as a number; I defined a num() function for that;
better than static_cast<unsigned int>, because its occurrences are
easier to find in the codebase.

Change includes fixing several places, e.g. messages, where unsigned int/short was
used instead of MacNodeId.