refactoring: port iterations over containers to use range-based 'for' loops and structured binding, using LLM
...where appropriate. (E.g. loops that delete from the iterated
container cannot be ported).

Using the following command, and cherry-picking the changes that were
good "as is":

llmtool --save-prompt --blocks-containing '\.begin\(\)' --without-context . --prompt "Try to convert for loops that iterate over STL containers to range-based 'for' loops instead. When iterating over a map, use structured bindings (and try to find good names for the key and value variables). DO NOT CHANGE ANYTHING ELSE."