Pdcp: move setIds() into the analyzePacket() (from entity into main Pdcp module)
Rationale: determining source/dest nodeIds for packets is NOT the job of
the PDCP protocol, and such code is already collected in analyzePacket()
which is going to be factored out into a separate module (or at least, into
another module) later.

This temporarily results in code duplication; we'll deal with that during
a future refactoring.