Pdcp: do not tweak srcId/destId in FlowControlInfo when sending DL packet over the X2 link
For DualConnectivity / Split Bearer, all DL packets arrive at the Master,
and those that go via the Secondary Node must be tunnelled over to it
via X2. Existing code tweaked the FlowControlInfo srcId/destId field to make it
look like the packet goes Enb-to-Enb, but that is not necessary.
This patch gets rid of that tweaking.

Note: the useNR flag selects whether the packet should go via the Master
or Secondary Enb.